feat(bundle)!: rename yfm ––> markdown in variable and file names
