+++ demo/editor-in-editor/index.tsx
@@ -3,7 +3,7 @@ import React, {useEffect} from 'react';
 import {toaster} from '@gravity-ui/uikit/toaster-singleton-react-18';
 
 import {BaseNode} from '../../src';
-import {YfmEditorView, useYfmEditor} from '../../src/bundle';
+import {MarkdownEditorView, useMarkdownEditor} from '../../src/bundle';
 import {VERSION} from '../../src/version';
 import {block} from '../cn';
 
@@ -16,7 +16,7 @@ import {
 const b = block('playground');
 
 export const PlaygroundEditorInEditor: React.FC = () => {
-    const mdEditor = useYfmEditor({
+    const mdEditor = useMarkdownEditor({
         initialEditorType: 'wysiwyg',
         initialToolbarVisible: true,
         allowHTML: false,
@@ -52,7 +52,7 @@ export const PlaygroundEditorInEditor: React.FC = () => {
             </div>
             <hr />
             <div className={b('editor')}>
-                <YfmEditorView
+                <MarkdownEditorView
                     stickyToolbar
                     settingsVisible
                     editor={mdEditor}
