+++ src/bundle/index.ts
@@ -1,8 +1,8 @@
 export type {ExtensionsOptions} from './wysiwyg-preset';
-export type {Editor, EditorType as YfmEditorType, RenderPreview, SplitMode} from './Editor';
+export type {Editor, EditorType as MarkdownEditorType, RenderPreview, SplitMode} from './Editor';
 export * from './context';
-export * from './useYfmEditor';
-export * from './YfmEditorView';
+export * from './useMarkdownEditor';
+export * from './MarkdownEditorView';
 
 // For project that do not support star export (export * as something from '...')
 import * as markupToolbarConfigs from './config/markup';
